refactor: simplified the import for types
