+++ jobs/Frontend/mews-movie-app-kayode-olumo/types.ts
@@ -1,0 +1,1 @@
+export * from "./lib/types/tmdb";
